requirements.txt
numpy
pandas
scikit-learn
tensorflow
fastapi
uvicorn
streamlit
openai
joblib
requests
python-dotenv
seaborn
matplotlib
imblearn